Split runs.tsx list/detail into separate files (400 → 2 line shim)
- runs.tsx                shim — re-exports both named pages so the
                          App.tsx import path stays stable
- runs/RunsListPage.tsx   list of all flow runs with status badges
- runs/RunDetailPage.tsx  detail view with stages, approve/reject UI,
                          shared-context (blackboard) viewer
- runs/StageCard.tsx      single stage card with icon, status badge,
                          notes, error, collapsible output (markdown)
- runs/styles.ts          RUN_STATUS_STYLE, STAGE_ICON, STAGE_TONE,
                          prettifyStageId — shared between both pages

The two pages were never really sharing component logic — just the
status style maps. Now each is independently editable.
